Add files via upload
Couldn't figure out how to make sure the work for assignment 3 and 4 stayed separate, so I made a new branch specifically for the additions for assignment 4. I apologize for the lateness of this portion. I just wanted to make sure the Maven version was easy to tell apart from the old version labeling-wise.